Move CLI code to the common module
This allows the CLI argument parsing code to be unit tested.

Add unit tests for most of the argument corner cases.
